apps/api/requirements.txt
# ATTREQ Backend Dependencies
# Latest stable versions as of October 2025

# Core FastAPI framework
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.35
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Data validation and settings
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.9

# HTTP client
httpx>=0.27.0

# Async file I/O
aiofiles>=23.2.0

# Environment management
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Linting and formatting
ruff>=0.8.0

# AI/ML dependencies (Phase 3)
rembg>=2.0.55
onnxruntime>=1.16.0
Pillow>=10.0.0
weaviate-client>=4.0.0
redis>=5.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
